Refactor HivePrestoToVeloxConnector to standalone file
