+++ presto-native-execution/presto_cpp/main/connectors/PrestoToVeloxConnector.h
@@ -35,26 +35,6 @@ void unregisterPrestoToVeloxConnector(const std::string& connectorName);
 const PrestoToVeloxConnector& getPrestoToVeloxConnector(
     const std::string& connectorName);
 
-std::vector<velox::common::Subfield> toRequiredSubfields(
-    const protocol::List<protocol::Subfield>& subfields);
-
-velox::connector::hive::HiveColumnHandle::ColumnType toHiveColumnType(
-    protocol::hive::ColumnType type);
-
-std::unique_ptr<velox::connector::ConnectorTableHandle> toHiveTableHandle(
-    const protocol::TupleDomain<protocol::Subfield>& domainPredicate,
-    const std::shared_ptr<protocol::RowExpression>& remainingPredicate,
-    bool isPushdownFilterEnabled,
-    const std::string& tableName,
-    const protocol::List<protocol::Column>& dataColumns,
-    const protocol::TableHandle& tableHandle,
-    const protocol::Map<protocol::String, protocol::String>& tableParameters,
-    const VeloxExprConverter& exprConverter,
-    const TypeParser& typeParser);
-
-velox::common::CompressionKind toFileCompressionKind(
-    const protocol::hive::HiveCompressionCodec& hiveCompressionCodec);
-
 class PrestoToVeloxConnector {
  public:
   virtual ~PrestoToVeloxConnector() = default;
@@ -134,56 +114,6 @@ class PrestoToVeloxConnector {
   explicit PrestoToVeloxConnector(std::string connectorName)
       : connectorName_(std::move(connectorName)) {}
   const std::string connectorName_;
-};
-
-class HivePrestoToVeloxConnector final : public PrestoToVeloxConnector {
- public:
-  explicit HivePrestoToVeloxConnector(std::string connectorName)
-      : PrestoToVeloxConnector(std::move(connectorName)) {}
-
-  std::unique_ptr<velox::connector::ConnectorSplit> toVeloxSplit(
-      const protocol::ConnectorId& catalogId,
-      const protocol::ConnectorSplit* connectorSplit,
-      const protocol::SplitContext* splitContext) const final;
-
-  std::unique_ptr<velox::connector::ColumnHandle> toVeloxColumnHandle(
-      const protocol::ColumnHandle* column,
-      const TypeParser& typeParser) const final;
-
-  std::unique_ptr<velox::connector::ConnectorTableHandle> toVeloxTableHandle(
-      const protocol::TableHandle& tableHandle,
-      const VeloxExprConverter& exprConverter,
-      const TypeParser& typeParser,
-      velox::connector::ColumnHandleMap& assignments)
-      const final;
-
-  std::unique_ptr<velox::connector::ConnectorInsertTableHandle>
-  toVeloxInsertTableHandle(
-      const protocol::CreateHandle* createHandle,
-      const TypeParser& typeParser) const final;
-
-  std::unique_ptr<velox::connector::ConnectorInsertTableHandle>
-  toVeloxInsertTableHandle(
-      const protocol::InsertHandle* insertHandle,
-      const TypeParser& typeParser) const final;
-
-  std::unique_ptr<velox::core::PartitionFunctionSpec>
-  createVeloxPartitionFunctionSpec(
-      const protocol::ConnectorPartitioningHandle* partitioningHandle,
-      const std::vector<int>& bucketToPartition,
-      const std::vector<velox::column_index_t>& channels,
-      const std::vector<velox::VectorPtr>& constValues,
-      bool& effectivelyGather) const final;
-
-  std::unique_ptr<protocol::ConnectorProtocol> createConnectorProtocol()
-      const final;
-
- private:
-  std::vector<std::shared_ptr<const velox::connector::hive::HiveColumnHandle>>
-  toHiveColumns(
-      const protocol::List<protocol::hive::HiveColumnHandle>& inputColumns,
-      const TypeParser& typeParser,
-      bool& hasPartitionColumn) const;
 };
 
 class TpchPrestoToVeloxConnector final : public PrestoToVeloxConnector {
