+++ presto-native-execution/presto_cpp/main/connectors/Registration.cpp
@@ -13,6 +13,7 @@
  */
 #include "presto_cpp/main/connectors/Registration.h"
 #include "presto_cpp/main/connectors/IcebergPrestoToVeloxConnector.h"
+#include "presto_cpp/main/connectors/HivePrestoToVeloxConnector.h"
 #include "presto_cpp/main/connectors/SystemConnector.h"
 
 #ifdef PRESTO_ENABLE_ARROW_FLIGHT_CONNECTOR
